added tests for services and user create endpoint

+++ users/services.py
@@ -3,31 +3,43 @@ from . selectors import user_get
 from django.db import transaction
 from rest_framework.exceptions import ValidationError
 
-@transaction.atomic
-def user_create(*, email: str, password: str) -> User:
 
-    user: User = User.objects.create_user(email=email, password=password)
+def user_create(*, email: str, password: str, first_name: str, last_name) -> User:
+
+    user: User = User.objects.create_user(email=email, password=password, first_name=first_name, last_name=last_name)
 
     return user
 
-@transaction.atomic
-def user_update(id: int, data: dict[str, str | bool]) -> User:
-    user = user_get(id=id)
-    fields: list[str] = ['email', 'first_name', 'last_name', 'is_active', 'is_verified'] 
 
+def user_update(user_id: int, data: dict[str, str | bool]) -> User | None:
+    user = user_get(user_id=user_id)
+    if user is None:
+        return None
+    
+    fields: list[str] = ['email', 'first_name', 'last_name', 'is_active', 'is_verified'] 
     for record in data:
         if record in fields:
             setattr(user, record, data[record])
         else:
             # TODO add logger , validation error shouldn't give to much info in production 
-            raise ValidationError({'detail':f"Field '{record}' is not meant to be updated"})
-        
+            # raise ValidationError({'detail': f"Field '{record}' is not meant to be updated"})
+            return None
+
     user.full_clean()
     user.save()
     return user
         
-
-def user_delete(id: int):
-    user = user_get(id=id)
+def user_delete(user_id: int) -> dict[int | str] | None:
+    user = user_get(user_id=user_id)
+    if user is None:
+        return None
+    
+    user_data = {
+        "id": user.id,
+        "email": user.email,
+        "first_name": user.first_name,
+        "last_name": user.last_name,
+    }
 
     user.delete()
+    return user_data
